creating interface for getStockItemConfiguration, Removing GetLegacyStockItem Service
