--- app/code/Magento/InventoryCatalog/Model/GetLegacyStockItem.php
@@ -1,69 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-declare(strict_types=1);
-
-namespace Magento\InventoryCatalog\Model;
-
-use Magento\CatalogInventory\Api\Data\StockItemInterface;
-use Magento\CatalogInventory\Api\StockItemCriteriaInterfaceFactory;
-use Magento\CatalogInventory\Model\Stock\Item;
-use Magento\Framework\Exception\LocalizedException;
-use Magento\CatalogInventory\Model\Stock\StockItemRepository;
-
-/**
- * Service GetLegacyStockItem returns old inventory item data
- */
-class GetLegacyStockItem
-{
-    /**
-     * @var StockItemCriteriaInterfaceFactory
-     */
-    private $stockItemCriteriaFactory;
-
-    /**
-     * @var GetProductIdsBySkusInterface
-     */
-    private $getProductIdsBySkus;
-
-    /**
-     * @var StockItemRepository
-     */
-    private $stockItemRepository;
-    /**
-     * @var Item[]
-     */
-    private $item;
-
-    /**
-     * GetLegacyStockItem constructor.
-     * @param StockItemCriteriaInterfaceFactory $stockItemCriteriaFactory
-     * @param StockItemRepository $stockItemRepository
-     * @param GetProductIdsBySkusInterface $getProductIdsBySkus
-     */
-    public function __construct(
-        StockItemCriteriaInterfaceFactory $stockItemCriteriaFactory,
-        StockItemRepository $stockItemRepository,
-        GetProductIdsBySkusInterface $getProductIdsBySkus
-    ) {
-        $this->stockItemCriteriaFactory = $stockItemCriteriaFactory;
-        $this->stockItemRepository = $stockItemRepository;
-        $this->getProductIdsBySkus = $getProductIdsBySkus;
-    }
-
-    /**
-     * @param string $sku
-     * @return StockItemInterface|Item[]
-     */
-    public function execute(string $sku)
-    {
-        $productId = $this->getProductIdsBySkus->execute([$sku])[$sku];
-        $searchCriteria = $this->stockItemCriteriaFactory->create();
-        $searchCriteria->addFilter(StockItemInterface::PRODUCT_ID, StockItemInterface::PRODUCT_ID, $productId);
-        $legacyStockItem = $this->stockItemRepository->getList($searchCriteria);
-
-        return $legacyStockItem->getItems()[0];
-    }
-}
