+++ app/code/Magento/InventoryConfiguration/etc/di.xml
@@ -25,7 +25,7 @@
                 </item>
                 <item name="min_qty" xsi:type="array">
                     <item name="sort_order" xsi:type="number">40</item>
-                    <item name="condition" xsi:type="object">Magento\InventoryConfiguration\Model\ConfigMinQty</item>
+                    <item name="condition" xsi:type="object">Magento\InventorySales\Model\ConfigMinQty</item>
                 </item>
                 <item name="is_salable" xsi:type="array">
                     <item name="sort_order" xsi:type="number">30</item>
@@ -38,4 +38,5 @@
             </argument>
         </arguments>
     </type>
+    <preference for="Magento\InventoryConfigurationApi\Api\GetStockItemConfigurationInterface" type="Magento\InventoryConfiguration\Model\GetStockItemConfiguration"/>
 </config>
